+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+
+develop
+ - Started refracturing structure for auditory scales
+ -
+
 0.67 -> 0.68
  - Reading and Writing Audiofiles is now handled using the soundfile library
  - Fixed some tests and depricated numpy apis